Compare an integer argument against its container's size before the cast, and bound array sizes, repeat counts, and JSON indents, so a huge value is out of range on every platform

+++ include/barescript/value.h
@@ -321,7 +321,7 @@ bool bsObjectIterSorted(BSValue value, BSObjectIterFn iter, void *data);
  */
 
 /* Round a number to "digits" decimal digits */
-double bsNumberRound(double value, int digits);
+double bsNumberRound(double value, double digits);
 
 /* Parse a number string; returns false if parsing fails */
 bool bsNumberParse(const char *text, size_t size, double *result);
